Move example to package
