+++ Makefile
@@ -6,11 +6,8 @@ help:
 install: package.json ## install dependencies
 	@npm install
 
-run: example_install ## run the example
-	@cd example && ./node_modules/.bin/webpack-dev-server --hot --inline --config ./webpack.config.js
-
-example_install: example/package.json
-	@cd example && npm install
+run: ## run the example
+	@cd packages/ra-example && ./node_modules/.bin/webpack-dev-server --hot --inline --config ./webpack.config.js
 
 clean:
 	@rm -rf lib
