--- example/i18n/index.js
@@ -1,2 +1,0 @@
-export en from './en';
-export fr from './fr';
