--- example/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Admin on REST</title>
-    <style>
-    body {
-        margin: 0;
-    }
-    a {
-        text-decoration: none;
-        color: #00bcd4;
-    }
-    </style>
-    <meta name=viewport content="width=device-width, initial-scale=1">
-  </head>
-  <body>
-    <div id="root">
-    </div>
-    <!-- This script adds the Roboto font to our project. For more detail go to this site:  http://www.google.com/fonts#UsePlace:use/Collection:Roboto:400,300,500 -->
-    <script>
-      var WebFontConfig = {
-        google: { families: [ 'Roboto:300,400,500,700:latin' ] }
-      };
-      (function() {
-        var wf = document.createElement('script');
-        wf.src = 'https://ajax.googleapis.com/ajax/libs/webfont/1/webfont.js';
-        wf.type = 'text/javascript';
-        wf.async = 'true';
-        var s = document.getElementsByTagName('script')[0];
-        s.parentNode.insertBefore(wf, s);
-      })();
-    </script>
-    <script src="/static/bundle.js"></script>
-  </body>
-</html>
