+++ packages/ra-example/.babelrc
@@ -1,0 +1,6 @@
+{
+    "presets": ["es2015", "react", "stage-0"],
+    "plugins": [
+      "transform-react-jsx",
+    ]
+}
