+++ packages/ra-example/data.js
@@ -1,0 +1,424 @@
+export default {
+    posts: [
+        {
+            id: 1,
+            title:
+                'Accusantium qui nihil voluptatum quia voluptas maxime ab similique',
+            teaser:
+                'In facilis aut aut odit hic doloribus. Fugit possimus perspiciatis sit molestias in. Sunt dignissimos sed quis at vitae veniam amet. Sint sunt perspiciatis quis doloribus aperiam numquam consequatur et. Blanditiis aut earum incidunt eos magnam et voluptatem. Minima iure voluptatum autem. At eaque sit aperiam minima aut in illum.',
+            body:
+                '<p>Rerum velit quos est <strong>similique</strong>. Consectetur tempora eos ullam velit nobis sit debitis. Magni explicabo omnis delectus labore vel recusandae.</p><p>Aut a minus laboriosam harum placeat quas minima fuga. Quos nulla fuga quam officia tempore. Rerum occaecati ut eum et tempore. Nam ab repudiandae et nemo praesentium.</p><p>Cumque corporis officia occaecati ducimus sequi laborum omnis ut. Nam aspernatur veniam fugit. Nihil eum libero ea dolorum ducimus impedit sed. Quidem inventore porro corporis debitis eum in. Nesciunt unde est est qui nulla. Esse sunt placeat molestiae molestiae sed quia. Sunt qui quidem quos velit reprehenderit quos blanditiis ducimus. Sint et molestiae maxime ut consequatur minima. Quaerat rem voluptates voluptatem quos. Corporis perferendis in provident iure. Commodi odit exercitationem excepturi et deserunt qui.</p><p>Optio iste necessitatibus velit non. Neque sed occaecati culpa porro culpa. Quia quam in molestias ratione et necessitatibus consequatur. Est est tempora consequatur voluptatem vel. Mollitia tenetur non quis omnis perspiciatis deserunt sed necessitatibus. Ad rerum reiciendis sunt aspernatur.</p><p>Est ullam ut magni aspernatur. Eum et sed tempore modi.</p><p>Earum aperiam sit neque quo laborum suscipit unde. Expedita nostrum itaque non non adipisci. Ut delectus quis delectus est at sint. Iste hic qui ea eaque eaque sed id. Hic placeat rerum numquam id velit deleniti voluptatem. Illum adipisci voluptas adipisci ut alias. Earum exercitationem iste quidem eveniet aliquid hic reiciendis. Exercitationem est sunt in minima consequuntur. Aut quaerat libero dolorem.</p>',
+            views: 143,
+            average_note: 2.72198,
+            commentable: true,
+            pictures: {
+                first: {
+                    name: 'the picture name',
+                    url: 'http://www.photo-libre.fr/paysage/1.jpg',
+                    metas: {
+                        title: 'This is a great photo',
+                        definitions: ['72', '300'],
+                        authors: [
+                            {
+                                name: 'Paul',
+                                email: 'paul@email.com',
+                            },
+                            {
+                                name: 'Joe',
+                                email: 'joe@email.com',
+                            },
+                        ],
+                    },
+                },
+                second: {
+                    name: 'better name',
+                    url: 'http://www.photo-libre.fr/paysage/2.jpg',
+                },
+            },
+            published_at: new Date('2012-08-06'),
+            tags: [1, 3],
+            category: 'tech',
+            subcategory: 'computers',
+            backlinks: [
+                {
+                    date: '2012-08-09T00:00:00.000Z',
+                    url: 'http://example.com/bar/baz.html',
+                },
+            ],
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 2,
+            title: 'Sint dignissimos in architecto aut',
+            teaser:
+                'Quam earum itaque corrupti labore quas nihil sed. Dolores sunt culpa voluptates exercitationem eveniet totam rerum. Molestias perspiciatis rem numquam accusamus.',
+            body:
+                '<p>Aliquam magni <em>tempora</em> quas enim. Perspiciatis libero corporis sunt eum nam. Molestias est sunt molestiae natus.</p><p>Blanditiis dignissimos autem culpa itaque. Explicabo perferendis ullam officia ut quia nemo. Eaque perspiciatis perspiciatis est hic non ullam et. Expedita exercitationem enim sit ut dolore.</p><h2>Sed in sunt officia blanditiis ipsam maiores perspiciatis amet</h2><p>Vero fugiat facere officiis aut quis rerum velit. Autem eius sint ullam. Nemo sunt molestiae nulla accusantium est voluptatem voluptas sed. In blanditiis neque libero voluptatem praesentium occaecati nulla libero. Perspiciatis eos voluptatem facere voluptatibus. Explicabo quo eveniet nihil culpa. Qui eos officia consequuntur sed esse praesentium dolorum. Eius perferendis qui quia autem nostrum sed. Illum in ex excepturi voluptas. Qui veniam sit alias delectus nihil. Impedit est ut alias illum repellendus qui.</p><p>Veniam est aperiam quisquam soluta. Magni blanditiis praesentium sed similique velit ipsam consequatur. Porro omnis magni sunt incidunt aspernatur ut.</p>',
+            views: 563,
+            average_note: 3.48121,
+            commentable: true,
+            published_at: new Date('2012-08-08'),
+            tags: [3, 5],
+            category: 'lifestyle',
+            backlinks: [],
+            notifications: [],
+        },
+        {
+            id: 3,
+            title: 'Perspiciatis adipisci vero qui ipsam iure porro',
+            teaser:
+                'Ut ad consequatur esse illum. Ex dolore porro et ut sit. Commodi qui sed et voluptatibus laudantium.',
+            body:
+                '<p>Voluptatibus fugit sit praesentium voluptas vero vel. Reprehenderit quam cupiditate deleniti ipsum nisi qui. Molestiae modi sequi vel quibusdam est aliquid doloribus. Necessitatibus et excepturi alias necessitatibus magnam ea.</p><p>Dolor illum dolores qui et pariatur inventore incidunt molestias. Exercitationem ipsum voluptatibus voluptatum velit sint vel qui. Odit mollitia minus vitae impedit voluptatem. Voluptas ullam temporibus inventore fugiat pariatur odit molestias.</p><p>Atque est qui alias eum. Quibusdam rem ut dolores voluptate totam. Sit cumque perferendis sed a iusto laudantium quae et. Voluptatibus vitae natus quia laboriosam et deserunt. Doloribus fuga aut quo tempora animi eaque consequatur laboriosam.</p>',
+            views: 467,
+            commentable: true,
+            published_at: new Date('2012-08-08'),
+            tags: [1, 2],
+            category: 'tech',
+            backlinks: [
+                {
+                    date: '2012-08-10T00:00:00.000Z',
+                    url: 'http://example.com/foo/bar.html',
+                },
+                {
+                    date: '2012-08-14T00:00:00.000Z',
+                    url: 'https://blog.johndoe.com/2012/08/12/foobar.html',
+                },
+                {
+                    date: '2012-08-22T00:00:00.000Z',
+                    url: 'https://foo.bar.com/lorem/ipsum',
+                },
+                {
+                    date: '2012-08-29T00:00:00.000Z',
+                    url: 'http://dicta.es/nam_doloremque',
+                },
+            ],
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 4,
+            title: 'Maiores et itaque aut perspiciatis',
+            teaser:
+                'Et quo voluptas odit veniam omnis dolores. Odit commodi consequuntur necessitatibus dolorem officia. Reiciendis quas exercitationem libero sed. Itaque non facilis sit tempore aut doloribus.',
+            body:
+                '<p>Sunt sunt aut est et consequatur ea dolores. Voluptatem rerum cupiditate dolore. Voluptas sit sapiente corrupti error ducimus. Qui enim aut possimus qui. Impedit voluptatem sed inventore iusto et ut et. Maxime sunt qui adipisci expedita quisquam. Velit ea ut in blanditiis eos doloribus.</p><p>Qui optio ad magnam eius. Est id velit ratione eum corrupti non vitae. Quam consequatur animi sed corrupti quae sed deserunt. Accusamus eius eos recusandae eum quia id.</p><p>Voluptas omnis omnis culpa est vel eum. Ut in tempore harum voluptates odit delectus sit et. Consequuntur quod nihil veniam natus placeat provident. Totam ut fuga vitae in. Possimus cumque quae voluptatem asperiores vitae officiis dolores. Qui autem eos dolores eius. Iure ut delectus quis voluptatem. Velit at incidunt minus laboriosam culpa. Pariatur ipsa ut enim dolor. Sed magni sunt molestiae voluptas ut illum. Sit consequuntur laborum aliquid delectus in. Consectetur dicta asperiores itaque aut mollitia. Minus praesentium officiis voluptas a officiis ad beatae.</p>',
+            views: 685,
+            average_note: 1.2319,
+            commentable: false,
+            published_at: new Date('2012-08-12'),
+            tags: [],
+            category: 'lifestyle',
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 5,
+            title: 'Sed quo et et fugiat modi',
+            teaser:
+                'Consequuntur id aut soluta aspernatur sit. Aut doloremque recusandae sit saepe ut quas earum. Quae pariatur iure et ducimus non. Cupiditate dolorem itaque in sit.',
+            body:
+                '<p>Aut molestiae quae explicabo voluptas. Assumenda ea ipsam quia. Rerum rerum magnam sunt doloremque dolorem nulla. Eveniet ut aliquam est dignissimos nisi molestias dicta. Dolorum et id esse illum. Ea omnis nesciunt tempore et aut. Ut ullam totam doloribus recusandae est natus voluptatum officiis. Ea quam eos velit ipsam non accusamus praesentium.</p><p>Animi et minima alias sint. Reiciendis qui ipsam autem fugit consequuntur veniam. Vel cupiditate voluptas enim dolore cum ad. Ut iusto eius et.</p><p>Quis praesentium aut aut aut voluptas et. Quam laudantium at laudantium amet. Earum quidem eos earum quaerat nihil libero quia sed.</p><p>Autem voluptatem nostrum ullam numquam quis. Et aut unde nesciunt officiis nam eos ut distinctio. Animi est explicabo voluptas officia quos necessitatibus. Omnis debitis unde et qui rerum. Nisi repudiandae autem mollitia dolorum veritatis aut. Rem temporibus labore repellendus enim consequuntur dicta autem. Illum illo inventore possimus officiis quidem.</p><p>Ullam accusantium eaque perspiciatis. Quidem dolor minus aut quidem. Praesentium earum beatae eos eligendi nostrum. Dolor nam quo aut.</p><p>Accusamus aut tempora omnis magni sit quos eos aut. Vitae ut inventore facere neque rerum. Qui esse rem cupiditate sit.</p><p>Est minus odio sint reprehenderit. Consectetur dolores eligendi et quaerat sint vel magni. Voluptatum hic cum placeat ad ea reiciendis laborum et. Eos ab id suscipit.</p>',
+            views: 559,
+            average_note: 3,
+            commentable: true,
+            published_at: new Date('2012-08-05'),
+            category: 'tech',
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 6,
+            title: 'Minima ea vero omnis odit officiis aut',
+            teaser:
+                'Omnis rerum voluptatem illum. Amet totam minus id qui aspernatur. Adipisci commodi velit sapiente architecto et molestias. Maiores doloribus quis occaecati quidem laborum. Quae quia quaerat est itaque. Vero assumenda quia tempora libero dicta quis asperiores magnam. Necessitatibus accusantium saepe commodi ut.',
+            body:
+                '<p>Sit autem rerum inventore repellendus. Enim placeat est ea dolor voluptas nisi alias. Repellat quam laboriosam repudiandae illum similique omnis non exercitationem. Modi mollitia omnis sed vel et expedita fugiat. Esse laboriosam doloribus deleniti atque quidem praesentium aliquid. Error animi ab excepturi quia. Et voluptates voluptatem et est quibusdam aspernatur. Fugiat consequatur veritatis commodi enim quaerat sint. Quis quae fuga exercitationem dolorem enim laborum numquam. Iste necessitatibus repellat in ea nihil et rem. Corporis dolores sed vitae consectetur dolores qui dicta. Laudantium et suscipit odit quidem qui. Provident libero eveniet distinctio debitis odio cum id dolorum. Consequuntur laboriosam qui ut magni sit dicta. Distinctio fugit voluptatibus voluptatem suscipit incidunt ut cupiditate. Magni harum in aut alias veniam. Eos aut impedit ut et. Iure aliquid adipisci aliquam et ab et qui. Itaque quod consequuntur dolore asperiores architecto neque. Exercitationem eum voluptas ut quis hic quo. Omnis quas porro laudantium. Qui magnam et totam quibusdam in quo. Impedit laboriosam eum sint soluta facere ut voluptatem.</p>',
+            views: 208,
+            average_note: 3.1214,
+            published_at: new Date('2012-09-05'),
+            tags: [1, 4],
+            category: 'tech',
+            notifications: [42],
+        },
+        {
+            id: 7,
+            title: 'Illum veritatis corrupti exercitationem sed velit',
+            teaser:
+                'Omnis hic quo aperiam fugiat iure amet est. Molestias ratione aut et dolor earum magnam placeat. Ad a quam ea amet hic omnis rerum.',
+            body:
+                '<p>Omnis sunt maxime qui consequatur perspiciatis et dolor. Assumenda numquam sit rerum aut dolores. Repudiandae rerum et quisquam. Perferendis cupiditate sequi non similique eum accusamus voluptas.</p><p>Officiis in voluptatum culpa ut eaque laborum. Sit quos velit sed ad voluptates. Alias aut quo accusantium aut cumque perferendis. Numquam rerum vel et est delectus. Mollitia dolores voluptatum accusantium id rem. Autem dolorem similique earum. Deleniti qui iusto et vero. Enim quaerat ipsum omnis magni. Autem magnam vero nulla impedit distinctio. Sequi laudantium ut animi enim recusandae et voluptatum. Dicta architecto nostrum voluptas consequuntur ea. Porro odio illo praesentium qui. Quia sit sed labore porro. Minima odit nemo sint praesentium. Ea sapiente quis aut. Qui cumque aut repudiandae in. Ipsam mollitia ab vitae iusto maxime. Eaque qui impedit et ea dolor aut. Tenetur ut nihil sed. Eum doloremque harum ipsam vel eos ut enim.</p>',
+            views: 133,
+            average_note: null,
+            commentable: true,
+            published_at: new Date('2012-09-29'),
+            tags: [3, 4],
+            category: 'tech',
+            notifications: [12, 31],
+        },
+        {
+            id: 8,
+            title:
+                'Culpa possimus quibusdam nostrum enim tempore rerum odit excepturi',
+            teaser:
+                'Qui quos exercitationem itaque quia. Repellat libero ut recusandae quidem repudiandae ipsam laudantium. Eveniet quos et quo omnis aut commodi incidunt.',
+            body:
+                '<p>Laudantium voluptatem non facere officiis qui natus natus. Ex perspiciatis quia dolor earum. In rerum deleniti voluptas quo quia adipisci voluptatibus.</p><p>Mollitia eos quaerat ad. Et non aliquam velit. Doloremque repudiandae earum suscipit deleniti.</p><p>Debitis voluptatem possimus saepe. Rerum nam est neque voluptate quae ratione et quaerat. Fugiat et ullam adipisci numquam. Atque qui cum quae quod qui reprehenderit. Veritatis odio eligendi est odit minima ut dolores. Blanditiis aut rem aliquam nulla esse odit. Quibusdam quam natus eos tenetur nemo eligendi velit nam. Consequatur libero eius quia impedit neque fuga. Accusantium sunt accusantium eaque illum dicta. Expedita explicabo quia soluta.</p><p>Dolores aperiam rem velit id provident quo ea. Modi illum voluptate corrupti recusandae optio. Voluptatem architecto numquam reiciendis quo nostrum suscipit. Dolore repellat deleniti nihil omnis illum explicabo nihil. Alias maxime hic minus voluptas odio id dolorum. Neque perferendis repellendus autem consequatur consequatur doloribus. Sit aspernatur nisi aliquam rem voluptas occaecati.</p><p>In eveniet nostrum culpa totam officia doloremque. Fugiat maxime magni aut magnam praesentium vel facere. Tempora soluta possimus omnis modi et qui minus. Consequatur et suscipit autem quia nulla.</p><p>Qui eum aliquid inventore at. Qui provident perspiciatis sed eum eos sunt eveniet autem. Ducimus velit tenetur sed. Quas laboriosam dicta ipsa id fugiat. Hic nihil laboriosam atque natus. Quam natus esse est error molestiae nulla. Odit ut dolorem laborum quidem quis alias. Labore sint porro et reprehenderit ut dolorem vel dolorum. Dolores suscipit ut dolores possimus id dicta cupiditate. Est cum dolorum dolores ducimus quia reprehenderit. Iste suscipit molestias voluptatem molestiae. Nostrum modi dicta qui deleniti. Reprehenderit voluptatem soluta non in labore. Voluptatem ut illo illo harum voluptas cumque. Tempora illo distinctio qui aut.</p><p>Eaque voluptatem eos omnis qui dolor non possimus. Distinctio ratione facere doloremque rerum qui voluptas et. Cum incidunt numquam molestias et labore odio sunt aut. Aut pariatur dignissimos est atque.</p>',
+            views: 557,
+            average_note: null,
+            commentable: false,
+            published_at: new Date('2012-10-02'),
+            tags: [5, 1],
+            category: 'lifestyle',
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 9,
+            title: 'A voluptas eius eveniet ut commodi dolor',
+            teaser:
+                'Sed necessitatibus nesciunt nesciunt aut non sunt. Quam ut in a sed ducimus eos qui sint. Commodi illo necessitatibus sint explicabo maiores. Maxime voluptates sit distinctio quo excepturi. Qui aliquid debitis repellendus distinctio et aut. Ex debitis et quasi id.',
+            body:
+                '<p>Consequatur temporibus explicabo vel laudantium totam. Voluptates nihil numquam accusamus ut unde quo. Molestiae dolores quas sit aliquam. Sit et fuga necessitatibus natus fugit voluptas et. Esse vitae sed sit eius.</p><p>Accusantium aliquam accusamus illo eum. Excepturi molestiae et earum qui. Iste dolor eligendi est vero iure eos nesciunt. Qui aspernatur repellendus id rerum consequatur ut. Quis ab quos fugit dicta aut voluptas. Rerum aut esse dolor. Illo iste ullam possimus nam nam assumenda molestiae est.</p><p>In porro nesciunt cumque in sint vel architecto. Aliquam et in numquam quae explicabo. Deserunt suscipit sunt excepturi optio molestiae. Facilis saepe eaque commodi provident ad voluptates eligendi.</p><p>Magnam et neque ad sed qui laborum et. Aut dolorem maxime harum. Molestias aut facere vitae voluptatem.</p><p>Excepturi odit doloremque eos quisquam sunt. Veniam repudiandae nisi dolorum nam quos. Qui voluptatem enim enim. Dolorum eveniet eaque expedita est tempore. Expedita amet blanditiis esse qui. Nam dolor odio nihil nobis quas quia exercitationem. Iusto ut ut reiciendis sint laudantium et distinctio. Vitae architecto accusamus quos dolores laudantium doloribus alias. Est est esse autem repellat. Assumenda officia aperiam sequi facere distinctio ut. Magnam qui assumenda eligendi sint. Architecto autem harum qui ea quos ut nesciunt et. Optio quidem sit ex quos provident. Et dolor dicta et laudantium. Incidunt id quo enim atque molestiae quam repudiandae omnis. Sed nam voluptatem dolores natus quisquam. Sit nostrum voluptate sed asperiores. Saepe eaque et illum aperiam. Maxime tenetur sunt reiciendis.</p><p>Ducimus quia dolorem voluptas ea. Fuga eum architecto eius cum est quibusdam eligendi est. In ut aperiam ea ut.</p>',
+            views: 143,
+            average_note: 3.1214,
+            commentable: true,
+            published_at: new Date('2012-10-16'),
+            tags: [],
+            category: 'tech',
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 10,
+            title: 'Totam vel quasi a odio et nihil',
+            teaser:
+                'Excepturi veritatis velit rerum nemo voluptatem illum tempora eos. Et impedit sed qui et iusto. A alias asperiores quia quo.',
+            body:
+                '<p>Voluptas iure consequatur repudiandae quibusdam iure. Quibusdam consequatur sit cupiditate aut eum iure. Provident ut aut est itaque ut eligendi sunt.</p><p>Odio ipsa dolore rem occaecati voluptatum neque. Quia est minima totam est dicta aliquid sed. Doloribus ea eligendi qui odit. Consectetur aut illum aspernatur exercitationem ut. Distinctio sapiente doloribus beatae natus mollitia. Nostrum cum magni autem expedita natus est nulla totam.</p><p>Et possimus quia aliquam est molestiae eum. Dicta nostrum ea rerum omnis. Ut hic amet sequi commodi voluptatem ut. Nulla magni totam placeat asperiores error.</p>',
+            views: 721,
+            average_note: 4.121,
+            commentable: true,
+            published_at: new Date('2012-10-19'),
+            tags: [1, 4],
+            category: 'lifestyle',
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 11,
+            title: 'Omnis voluptate enim similique est possimus',
+            teaser:
+                'Velit eos vero reprehenderit ut assumenda saepe qui. Quasi aut laboriosam quas voluptate voluptatem. Et eos officia repudiandae quaerat. Mollitia libero numquam laborum eos.',
+            body:
+                '<p>Ut qui a quis culpa impedit. Harum quae sunt aspernatur dolorem minima et dolorum. Consequatur sunt eveniet sit perspiciatis fuga praesentium. Quam voluptatem a ullam accusantium debitis eum consectetur.</p><p>Voluptas rem impedit omnis maiores saepe. Eum consequatur ut et consequatur repellat. Quos dolorem dolorum nihil dolor sit optio velit. Quasi quaerat enim omnis ipsum.</p><p>Officia asperiores ut doloribus. Architecto iste quia illo non. Deleniti enim odio aut amet eveniet. Modi sint aut excepturi quisquam error sed officia. Nostrum enim repellendus inventore minus. Itaque vitae ipsam quasi. Qui provident vero ab facere. Sit enim provident doloremque minus quam. Voluptatem expedita est maiores nihil est voluptatem error. Asperiores ut a est ducimus hic optio. Natus omnis ullam consectetur ducimus nisi sint ducimus odit. Soluta cupiditate ipsam magnam.</p><p>Illum magni aut autem in sed iure. Ea explicabo ducimus officia corrupti ipsam minima minima. Nihil ab similique modi sunt unde nisi. Iusto quis iste ut aut earum magni. Nisi nisi minima sapiente quos aut libero maxime. Ut consequuntur sit vel odio suscipit fugiat tempore et. Et eveniet aut voluptatibus aliquid accusantium quis qui et. Veniam rem ut et. Vel officiis et voluptatum eaque ipsum sit. Sed iste rem ipsam dolor maiores. Et animi aspernatur aut error. Quisquam veritatis voluptatem magnam id. Blanditiis dolorem quo et voluptatum.</p>',
+            views: 294,
+            average_note: 3.12942,
+            commentable: true,
+            published_at: new Date('2012-10-22'),
+            tags: [4, 3],
+            category: 'tech',
+            subcategory: 'computers',
+            pictures: null,
+            backlinks: [
+                {
+                    date: '2012-10-29T00:00:00.000Z',
+                    url: 'http://dicta.es/similique_pariatur',
+                },
+            ],
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 12,
+            title: 'Qui tempore rerum et voluptates',
+            teaser:
+                'Occaecati rem perferendis dolor aut numquam cupiditate. At tenetur dolores pariatur et libero asperiores porro voluptas. Officiis corporis sed eos repellendus perferendis distinctio hic consequatur.',
+            body:
+                '<p>Praesentium corrupti minus molestias eveniet mollitia. Sit dolores est tenetur eos veritatis. Vero aut molestias provident ducimus odit optio.</p><p>Minima amet accusantium dolores et. Iste eos necessitatibus iure provident rerum repellendus reiciendis eos. Voluptate dolorem dolore aliquid sed maiores.</p><p>Ut quia excepturi quidem quidem. Cupiditate qui est rerum praesentium consequatur ad. Minima rem et est. Ut odio nostrum fugit laborum. Quis vitae occaecati tenetur earum non architecto.</p><p>Minima est nobis accusamus sunt explicabo fuga. Ut ut ut officia labore ratione animi saepe et.</p><p>Accusamus quae ex rerum est eos nesciunt et. Nemo nam consequatur earum necessitatibus et. Eum corporis corporis quia at nihil consectetur accusamus. Ea eveniet et culpa maxime.</p><p>Et et quisquam odio sapiente. Voluptas ducimus beatae ratione et soluta esse ut animi. Ipsa architecto veritatis cumque in.</p><p>Voluptatem dolore sint aliquam excepturi. Pariatur quisquam a eum. Aut et sit quis et dolorem omnis. Molestias id cupiditate error ab.</p><p>Odio ut deleniti incidunt vel dolores eligendi. Nemo aut commodi accusamus alias reprehenderit dolorum eaque. Iure fugit quis occaecati aspernatur tempora iste.</p><p>Omnis repellat et sequi numquam accusantium doloribus eum totam. Ab assumenda facere qui voluptate. Temporibus non ipsa officia. Corrupti omnis ut dolores velit aliquam ut omnis consequuntur.</p>',
+            views: 719,
+            average_note: 2,
+            commentable: true,
+            published_at: new Date('2012-11-07'),
+            tags: [],
+            category: 'lifestyle',
+            subcategory: 'fitness',
+            pictures: { first: {}, second: {} },
+            backlinks: [
+                {
+                    date: '2012-08-07T00:00:00.000Z',
+                    url: 'http://example.com/foo/bar.html',
+                },
+                {
+                    date: '2012-08-12T00:00:00.000Z',
+                    url: 'https://blog.johndoe.com/2012/08/12/foobar.html',
+                },
+            ],
+            notifications: [12, 31, 42],
+        },
+        {
+            id: 13,
+            title: 'Fusce massa lorem, pulvinar a posuere ut, accumsan ac nisi',
+            teaser:
+                'Quam earum itaque corrupti labore quas nihil sed. Dolores sunt culpa voluptates exercitationem eveniet totam rerum. Molestias perspiciatis rem numquam accusamus.',
+            body:
+                '<p>Curabitur eu odio ullamcorper, pretium sem at, blandit libero. Nulla sodales facilisis libero, eu gravida tellus ultrices nec. In ut gravida mi. Vivamus finibus tortor tempus egestas lacinia. Cras eu arcu nisl. Donec pretium dolor ipsum, eget feugiat urna iaculis ut.</p> <p>Nullam lacinia accumsan diam, ac faucibus velit maximus ac. Donec eros ligula, ullamcorper sit amet varius eget, molestie nec sapien. Donec ac est non tellus convallis condimentum. Aliquam non vehicula mauris, ac rhoncus mi. Integer consequat ipsum a posuere ornare. Quisque mollis finibus libero scelerisque dapibus. </p>',
+            views: 222,
+            average_note: 4,
+            commentable: true,
+            published_at: new Date('2012-12-01'),
+            tags: [3, 5],
+            category: 'lifestyle',
+            backlinks: [],
+            notifications: [],
+        },
+    ],
+    comments: [
+        {
+            id: 1,
+            author: {},
+            post_id: 6,
+            body:
+                "Queen, tossing her head through the wood. 'If it had lost something; and she felt sure it.",
+            created_at: new Date('2012-08-02'),
+        },
+        {
+            id: 2,
+            author: {
+                name: 'Kiley Pouros',
+                email: 'kiley@gmail.com',
+            },
+            post_id: 9,
+            body:
+                "White Rabbit: it was indeed: she was out of the ground--and I should frighten them out of its right paw round, 'lives a March Hare. 'Sixteenth,'.",
+            created_at: new Date('2012-08-08'),
+        },
+        {
+            id: 3,
+            author: {
+                name: 'Justina Hegmann',
+            },
+            post_id: 3,
+            body:
+                "I'm not Ada,' she said, 'and see whether it's marked \"poison\" or.",
+            created_at: new Date('2012-08-02'),
+        },
+        {
+            id: 4,
+            author: {
+                name: 'Ms. Brionna Smitham MD',
+            },
+            post_id: 6,
+            body:
+                "Dormouse. 'Fourteenth of March, I think I can say.' This was such a noise inside, no one else seemed inclined.",
+            created_at: new Date('2014-09-24'),
+        },
+        {
+            id: 5,
+            author: {
+                name: 'Edmond Schulist',
+            },
+            post_id: 1,
+            body:
+                "I ought to tell me your history, you know,' the Hatter and the happy summer days. THE.",
+            created_at: new Date('2012-08-07'),
+        },
+        {
+            id: 6,
+            author: {
+                name: 'Danny Greenholt',
+            },
+            post_id: 6,
+            body:
+                'Duchess asked, with another hedgehog, which seemed to be lost: away went Alice after it, never once considering how in the other. In the very tones of.',
+            created_at: new Date('2012-08-09'),
+        },
+        {
+            id: 7,
+            author: {
+                name: 'Luciano Berge',
+            },
+            post_id: 5,
+            body:
+                "While the Panther were sharing a pie--' [later editions continued as follows.",
+            created_at: new Date('2012-09-06'),
+        },
+        {
+            id: 8,
+            author: {
+                name: 'Annamarie Mayer',
+            },
+            post_id: 5,
+            body:
+                "I tell you, you coward!' and at once and put it more clearly,' Alice.",
+            created_at: new Date('2012-10-03'),
+        },
+        {
+            id: 9,
+            author: {
+                name: 'Breanna Gibson',
+            },
+            post_id: 2,
+            body:
+                "THAT. Then again--\"BEFORE SHE HAD THIS FIT--\" you never tasted an egg!' 'I HAVE tasted eggs, certainly,' said Alice, as she spoke. Alice did not like to have it.",
+            created_at: new Date('2012-11-06'),
+        },
+        {
+            id: 10,
+            author: {
+                name: 'Logan Schowalter',
+            },
+            post_id: 3,
+            body:
+                "I'd been the whiting,' said the Hatter, it woke up again with a T!' said the Gryphon. '--you advance twice--' 'Each with a growl, And concluded the banquet--] 'What IS the fun?' said.",
+            created_at: new Date('2012-12-07'),
+        },
+        {
+            id: 11,
+            author: {
+                name: 'Logan Schowalter',
+            },
+            post_id: 1,
+            body:
+                "I don't want to be?' it asked. 'Oh, I'm not Ada,' she said, 'and see whether it's marked \"poison\" or not'; for she had asked it aloud; and in despair she put her hand on the end of the.",
+            created_at: new Date('2012-08-05'),
+        },
+    ],
+    tags: [
+        {
+            id: 1,
+            name: 'Sport',
+            published: 1,
+        },
+        {
+            id: 2,
+            name: 'Technology',
+            published: false,
+        },
+        {
+            id: 3,
+            name: 'Code',
+            published: true,
+        },
+        {
+            id: 4,
+            name: 'Photo',
+            published: false,
+        },
+        {
+            id: 5,
+            name: 'Music',
+            published: 1,
+        },
+    ],
+    users: [
+        {
+            id: 1,
+            name: 'Logan Schowalter',
+            role: 'admin',
+        },
+        {
+            id: 2,
+            name: 'Breanna Gibson',
+            role: 'user',
+        },
+        {
+            id: 3,
+            name: 'Annamarie Mayer',
+            role: 'user',
+        },
+    ],
+};
