+++ packages/ra-example/i18n/index.js
@@ -1,0 +1,2 @@
+export en from './en';
+export fr from './fr';
