+++ packages/ra-example/webpack.config.js
@@ -1,0 +1,17 @@
+const path = require('path');
+
+module.exports = {
+    devtool: 'eval',
+    entry: './app.js',
+    output: {
+        path: path.join(__dirname, 'static'),
+        filename: 'bundle.js',
+        publicPath: '/static/',
+    },
+    module: {
+        loaders: [
+            { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader' },
+            { test: /\.css$/, loader: 'style-loader!css-loader' },
+        ],
+    },
+};
